# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API EDUCAFI",
        "description": "Documentación de la API EDUCAFI",
        "contact": {
            "email": "soporte@educafi.mx"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/APIeducafi/api/validar/buscar": {
            "post": {
                "tags": [
                    "Validar"
                ],
                "summary": "Buscar alumnos y profesores por criterios",
                "operationId": "941c223c3e56a7cc953206aefc050445",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido_paterno": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "apellido_materno": {
                                        "type": "string",
                                        "example": "García"
                                    },
                                    "correo_personal": {
                                        "type": "string",
                                        "example": "juan@mail.com"
                                    },
                                    "identificador": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resultados de la búsqueda"
                    },
                    "400": {
                        "description": "Debe enviar al menos un campo para buscar"
                    },
                    "500": {
                        "description": "Error al ejecutar la consulta"
                    }
                }
            }
        },
        "/APIeducafi/api/validar/verdatos": {
            "get": {
                "tags": [
                    "Validar"
                ],
                "summary": "Obtiene los datos validados de un usuario",
                "operationId": "67d34598e151353d3b1c1616e1934e7c",
                "parameters": [
                    {
                        "name": "identificador",
                        "in": "query",
                        "description": "Datos del profesor o Alumno validados",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tipo_usuario",
                        "in": "query",
                        "description": "Tipo de usuario: Profesor o Alumno",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Profesor",
                                "Alumno"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del usuario"
                    },
                    "400": {
                        "description": "Parámetros faltantes o inválidos"
                    },
                    "500": {
                        "description": "Error al ejecutar la función"
                    }
                }
            }
        }
    },
...
}